chore(Flow): extract delete dialog controller and subcomponents [#1752]
